refactor(scripts): extract corpus hub access and parse taxonomy
Slice 2 of #4830 needs to classify parse outcomes exactly the way the
slice-1 probe does, and to talk to the same dataset endpoints. Split those
two concerns out of probe_open_schematics.py so the manifest tooling shares
one implementation instead of growing a drifting copy:

- scripts/corpus/parse_taxonomy.py — outcome constants, format sniffing,
  exception classification, signature grouping, parse_payload/parse_file.
  Pure and offline (no network, no CLI), which is what makes it importable
  from tests/ where the network scripts are not.
- scripts/corpus/hf_hub.py — datasets-server access with retry/backoff,
  plus row_url() for pinning a record's fetch URL in a manifest.

Behaviour of the probe is unchanged; verified with the offline --dry-run
path.
